Split GUI folder scan actions
